+++ scancodeio/static/main.js
@@ -262,7 +262,7 @@ function displayOverlay() {
   const background = document.createElement("div");
   background.setAttribute("id", "background-overlay");
   background.className = "modal-background";
-  background.style.cssText = "z-index:100;color:white;text-align:center;padding-top:150px;position:fixed;";
+  background.style.cssText = "z-index:100;color:white;text-align:center;padding-top:150px;position:fixed;background-color:rgba(9, 10, 12, 0.86)";
   background.innerHTML = '<div class="fa-5x"><i class="fas fa-circle-notch fa-spin"></i></div>';
   document.body.appendChild(background);
   return background;
@@ -441,3 +441,98 @@ function displayPipelineStatusToast(run_status, pipeline_name, project_url) {
 
     bulmaToast.toast({...default_options, ...custom_options});
 }
+
+// Themes
+
+const STORAGE_KEY = "bulma-theme";
+const SYSTEM_THEME = "system";
+const DEFAULT_THEME = "light";
+
+const state = {
+  chosenTheme: SYSTEM_THEME, // light|dark|system
+  appliedTheme: DEFAULT_THEME, // light|dark
+  OSTheme: null, // light|dark|null
+};
+
+const $themeSwitchers = document.querySelectorAll(".js-themes a");
+
+const updateThemeUI = () => {
+  $themeSwitchers.forEach((el) => {
+    const swatchTheme = el.dataset.scheme;
+
+    if (state.chosenTheme === swatchTheme) {
+      el.classList.add("is-active");
+    } else {
+      el.classList.remove("is-active");
+    }
+  });
+};
+
+const setTheme = (theme, save = true) => {
+  state.chosenTheme = theme;
+  state.appliedTheme = theme;
+
+  if (theme === SYSTEM_THEME) {
+    state.appliedTheme = state.OSTheme;
+    document.documentElement.removeAttribute("data-theme");
+    window.localStorage.removeItem(STORAGE_KEY);
+  } else {
+    document.documentElement.setAttribute("data-theme", theme);
+
+    if (save) {
+      window.localStorage.setItem(STORAGE_KEY, theme);
+    }
+  }
+
+  updateThemeUI();
+};
+
+const toggleTheme = () => {
+  if (state.appliedTheme === "light") {
+    setTheme("dark");
+  } else {
+    setTheme("light");
+  }
+};
+
+const detectOSTheme = () => {
+  if (!window.matchMedia) {
+    // matchMedia method not supported
+    return DEFAULT_THEME;
+  }
+
+  if (window.matchMedia("(prefers-color-scheme: dark)").matches) {
+    // OS theme setting detected as dark
+    return "dark";
+  } else if (window.matchMedia("(prefers-color-scheme: light)").matches) {
+    return "light";
+  }
+
+  return DEFAULT_THEME;
+};
+
+// On load, check if any preference was saved
+const localTheme = window.localStorage.getItem(STORAGE_KEY);
+state.OSTheme = detectOSTheme();
+
+if (localTheme) {
+  setTheme(localTheme, false);
+} else {
+  setTheme(SYSTEM_THEME);
+}
+
+// Event listeners
+$themeSwitchers.forEach((el) => {
+  el.addEventListener("click", () => {
+    const theme = el.dataset.scheme;
+    setTheme(theme);
+  });
+});
+
+window
+  .matchMedia("(prefers-color-scheme: dark)")
+  .addEventListener("change", (event) => {
+    const theme = event.matches ? "dark" : "light";
+    state.OSTheme = theme;
+    setTheme(theme);
+  });